rename: data부분 api/로 변경
oauth에서 관리하던 data는 api 라우터로 바꿈

Related to: #42
